(Cosmetic) Made case type more consistent
The naming of the modules are in PascalCase, but the inside of the scripts are written in camelCase. Luau's official styleguide also states Luau should be written in PascalCase for consistency. Also, the underscore can hurt readability, someone would expect an underscore to be in-between two words, but you've used it here to prevent Roblox replacing the error message with the global "error" function.